requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pillow==12.0.0
sqlparse==0.5.3
